+++ R/london.R
@@ -2,21 +2,21 @@
 #'
 #' A classical test on problem solving.
 #'
-#' @templateVar .by TRUE
-#' @template params-template
+#' @template common
+#' @template options
 #' @return A [tibble][tibble::tibble-package] contains following values:
 #'   \item{total_score}{Total score defined .by the game itself.}
 #'   \item{mean_level}{Mean level reached.}
 #'   \item{level_score}{Sum of mean score (a ratio) for each level.}
 #' @export
-london <- function(data, .by = NULL) {
+london <- function(data, .by = NULL, .input = NULL, .extra = NULL) {
   .input <- list(
     name_level = "level",
     name_score = "score",
     name_outcome = "outcome",
     name_steps = "stepsused"
   ) |>
-    update_settings("preproc.input")
+    update_settings(.input)
   total_score <- data |>
     group_by(across(all_of(.by))) |>
     summarise(