+++ man-roxygen/options.R
@@ -1,0 +1,3 @@
+#' @param .input,.extra Each is a [list()] containing all the input variable
+#'   names and special values for certain variables. See more in the details
+#'   section.